feat: teleprompter screen

+++ app/src/main/java/com/scriptglance/data/model/presentation/PresentationActiveData.kt
@@ -4,6 +4,6 @@ data class PresentationActiveData(
     val currentReadingPosition: PresentationActiveCurrentReadingPosition,
     val structure: List<PresentationActiveStructureItem>,
     val currentPresentationStartDate: String?,
-    val currentOwnerUserId: Long,
+    val currentOwnerUserId: Int,
     val joinedUsers: List<PresentationActiveJoinedUser>
 )
